pyproject.toml
[tool.poetry]
name = "sp"
version = "0.1.0"
description = ""
authors = ["James King <jcking1034@gmail.com>"]

[[tool.poetry.source]]
name = "elfin"  # This name will be used in the configuration to retreive the proper credentials
url = "http://git.elfin.ucla/api/v4/projects/5/packages/pypi/simple"  # URL used to download your packages from

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
requests = "^2.27.1"
gunicorn = "^20.1.0"
SQLAlchemy = "^1.4.36"
elfin-ops-libs = "^3.1.16"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
